% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_package.R
\name{growth_estimator}
\alias{growth_estimator}
\title{The relationship between specific growth rate (G) and temperature in °C (T) 
where a,b,c,d and are parameters (input of the function).}
\usage{
growth_estimator(a = -0.497, b = 0.1656, c = 0.08588, d = -0.004266)
}
\description{
The relationship between specific growth rate (G) and temperature in °C (T) 
where a,b,c,d and are parameters (input of the function).
}
