% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_package.R
\name{summarize}
\alias{summarize}
\title{this is a function takes two tables as input. One is table with prices of fishes
and other is the table with number of fishes per location and a binary variable
for producing a graph. 
price_table = table with fish prices
number_table= table with number of fishes per location
graph= a variable with 1 for producing a plot and 0 otherwise. Defaults to 0}
\usage{
summarize(price_table, number_table, graph = 0)
}
\description{
this is a function takes two tables as input. One is table with prices of fishes
and other is the table with number of fishes per location and a binary variable
for producing a graph. 
price_table = table with fish prices
number_table= table with number of fishes per location
graph= a variable with 1 for producing a plot and 0 otherwise. Defaults to 0
}
